Now also handle boundary case of NaN and 0 for parseInt of donation amount. God, I hate javascript.

+++ app/helpers/store_helper.rb
@@ -1,7 +1,7 @@
 module StoreHelper
 
   def to_numeric(str)
-    str.blank? ? 0 : str.to_f
+    str.blank? ? 0 : str.to_i
   end
 
   # make a form field not be submitted, by removing its name attribute
